AI/img_forensic/requirements.txt
requests
json5
regex
python-dotenv
numpy
redis
piexif
pathlib
requests
Pillow
opencv-python-headless
PyWavelets
langchain_core
langchain_huggingface